requirements.txt
numpy~=1.18.1
torchvision~=0.5.0
chainer~=5.3.0
scipy~=1.4.1
scikit-image~=0.16.2
pillow~=7.0.0